+++ R/utils.R
@@ -1,2 +1,0 @@
-replaceMinus <- function(x)
-    gsub("-", "_", x)